Added support for RLE-based homref output
New option --homrefAsRLE allows homref positions to be output as run-length encoding.
The file format is inspired by the Wiggle format.
For each chromsome present in the BAM file, a header line is output in the homref RLE file:
chrom={}
where {} is the name of the chromsome.
Each entry of this file contains two values, a start position, and a run length giving the number of contiguous positions with homref genotype calls.
This is supported by the new data structure, homref_rle.